Feature complete GLB loading with tiny gltf (ktx, webp, and draco support)

+++ Source/Engine/Systems/Renderer/Core/Meshes/Mesh.h
@@ -9,13 +9,13 @@ namespace Engine
   struct Mesh
   {
     std::vector<Vertex> vertices; 
-    std::vector<uint16_t> indices; 
+    std::vector<uint32_t> indices; 
 
     std::shared_ptr<MeshBufferData> meshBufferData;
 
     Mesh() = default;
 
-    Mesh(std::vector<Vertex> v, std::vector<uint16_t> i)
+    Mesh(std::vector<Vertex> v, std::vector<uint32_t> i)
       : vertices(std::move(v)), indices(std::move(i))
     {}
   };
