+++ Source/Engine/Systems/Renderer/Core/Meshes/MeshBufferData.h
@@ -31,7 +31,7 @@ namespace Engine
 
 		GLuint GetIndexCount() const { return indexCount; }
 
-		void GenerateBuffersAndAABB(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices);
+		void GenerateBuffersAndAABB(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices);
 
 	};
