+++ Source/Engine/Systems/Renderer/Vulkan/VulkanIndexDraw.h
@@ -28,7 +28,7 @@ namespace Engine
 
 		void CreateMegaMeshBuffers(VkDeviceSize totalVertexBufferSize, VkDeviceSize totalIndexBufferSize);
 
-		void UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint16_t>& indices, MeshBufferData& meshData);
+		void UploadMeshToMegaBuffer(const std::vector<Vertex>& vertices, const std::vector<uint32_t>& indices, MeshBufferData& meshData);
 
 		void UpdateInstanceBuffer(uint32_t frameIndex);
